# backend/app/api/static/docs/openapi-3.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  "/v1/entities/{id}/offboard":
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Offboard an active asset
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.EntityOffboardRequest"
        description: Offboarding record
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOffboardingRecord"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  "/v1/entities/{id}/offboarding-history":
    get:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Get retained asset offboarding history, oldest recorded first
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.EntityOffboardingRecord"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
...
  "/v1/entities/{id}/reactivate":
    post:
      security:
        - Bearer: []
      tags:
        - Entities
      summary: Reactivate an offboarded asset without changing archive state
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            "*/*":
              schema:
                type: object
                additionalProperties:
                  type: string
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.EntityOffboardingRecord:
      type: object
      properties:
        createdAt:
          type: string
        customReason:
          type: string
        effectiveDate:
          type: string
        id:
          type: string
        notes:
          type: string
        outcome:
          type: string
        reactivatedAt:
          type: string
          nullable: true
...
    v1.EntityOffboardRequest:
      type: object
      properties:
        customReason:
          type: string
        effectiveDate:
          type: string
          example: 2026-10-09
        notes:
          type: string
        outcome:
          type: string
          enum:
            - sold
            - donated
            - disposed
            - recycled
            - lost
            - custom
...